Rename Main -> MainList and factor out MainBase
with account actions
