+++ Source/Scenes/Main/MainListViewController.swift
@@ -1,6 +1,4 @@
 /*
- *  MainViewController.swift
- *
  *  Copyright (C) 2018 Kenneth H. Cox
  *
  *  This program is free software; you can redistribute it and/or
@@ -30,8 +28,8 @@ struct ButtonAction {
     let handler: (() -> Void)
 }
 
-class MainViewController: UIViewController {
-    
+class MainListViewController: MainBaseViewController {
+
     //MARK: - fields
 
     @IBOutlet weak var accountButton: UIBarButtonItem!
@@ -63,17 +61,9 @@ class MainViewController: UIViewController {
             tableView.deselectRow(at: indexPath, animated: true)
         }
 
-        // cause func to be called when returning to app, e.g. from browser
-        NotificationCenter.default.addObserver(self, selector: #selector(applicationDidBecomeActive),
-                                               name: UIApplication.didBecomeActiveNotification, object: nil)
-
         self.fetchData()
     }
-    
-    override func viewWillDisappear(_ animated: Bool) {
-        NotificationCenter.default.removeObserver(self, name: UIApplication.didBecomeActiveNotification, object: nil)
-    }
-    
+
     //MARK: - Functions
     
     func setupButtons() {
@@ -207,63 +197,6 @@ class MainViewController: UIViewController {
         let unreadCount = messages.filter { $0.isPatronVisible && !$0.isDeleted && !$0.isRead }.count
         messagesButton.setBadge(text: (unreadCount > 0) ? String(unreadCount) : nil)
     }
-    
-    @objc func accountButtonPressed(sender: UIBarButtonItem) {
-        let haveMultipleAccounts = App.credentialManager.credentials.count > 1
-
-        // Create an action sheet to present the account options
-        let alertController = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
-        Style.styleAlertController(alertController)
-        
-        // Add an action for each stored account
-        if haveMultipleAccounts {
-            for credential in App.credentialManager.credentials {
-                let action = UIAlertAction(title: credential.chooserLabel, style: .default) { action in
-                    self.doSwitchAccount(toAccount: credential)
-                }
-                var imageName = "Account"
-                if credential.username == App.account?.username {
-                    action.isEnabled = false
-                    imageName = "Account with Checkmark"
-                }
-                if let icon = loadAssetImage(named: imageName) {
-                    action.setValue(icon, forKey: "image")
-                }
-                alertController.addAction(action)
-            }
-        }
-        
-        // Add remaining actions
-        alertController.addAction(UIAlertAction(title: "Add account", style: .default) { action in
-            self.doAddAccount()
-        })
-        alertController.addAction(UIAlertAction(title: "Logout", style: .destructive) { action in
-            self.doLogout()
-        })
-        alertController.addAction(UIAlertAction(title: "Cancel", style: .cancel))
-        
-        // iPad requires a popoverPresentationController
-        if let popoverController = alertController.popoverPresentationController {
-            popoverController.barButtonItem = sender
-        }
-
-        self.present(alertController, animated: true)
-    }
-    
-    func doSwitchAccount(toAccount storedAccount: Credential) {
-        App.switchCredential(credential: storedAccount)
-        self.popToLogin()
-    }
-    
-    func doAddAccount() {
-        App.switchCredential(credential: nil)
-        self.popToLogin(forAddingCredential: true)
-    }
-    
-    func doLogout() {
-        App.logout()
-        self.popToLogin()
-    }
 
     @IBAction func fullCatalogButtonPressed(_ sender: Any) {
         if let libraryURL = App.library?.url,
@@ -292,7 +225,7 @@ class MainViewController: UIViewController {
         }
     }
 
-    @objc func applicationDidBecomeActive() {
+    @objc override func applicationDidBecomeActive() {
         os_log("didBecomeActive: fetchData", log: log)
         fetchData()
     }
@@ -300,7 +233,7 @@ class MainViewController: UIViewController {
 }
 
 //MARK: - UITableViewDataSource
-extension MainViewController: UITableViewDataSource {
+extension MainListViewController: UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return buttons.count
@@ -333,7 +266,7 @@ extension MainViewController: UITableViewDataSource {
 }
 
 //MARK: - UITableViewDelegate
-extension MainViewController: UITableViewDelegate {
+extension MainListViewController: UITableViewDelegate {
 
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let action = buttons[indexPath.row]
